--- client/src/components/Professionals/ProSearch/ProList.js
@@ -1,4 +1,0 @@
-import React from 'react';
-export default function ProList(props) {
-    return <div className='wrapper'>{props.children}</div>;
-}
